refactor(cli): migrate land validators (bathymetry/fabdem/glaciers/soilgrids)
Move the four validator-only land backends to earthlens.<backend>.cli and drop
their _VALIDATORS entries + functions from core. Relocate bathymetry/soilgrids
unit tests to earthlens-land. Shrink the guard allow-list by four.
