+++ libs/core/tests/test_distribution_boundaries.py
@@ -1271,17 +1271,14 @@ class TestCliIsBackendAgnostic:
     PENDING = frozenset(
         {
             "asf",
-            "bathymetry",
             "catrare",
             "chc",
             "drought",
             "earthdata",
             "ecmwf",
             "eumetsat",
-            "fabdem",
             "gee",
             "ghsl",
-            "glaciers",
             "goes",
             "jaxa",
             "mswep",
@@ -1294,7 +1291,6 @@ class TestCliIsBackendAgnostic:
             "radklim",
             "s3",
             "sentinel_hub",
-            "soilgrids",
             "stac",
             "tropycal",
             "worldpop",
